requirements.txt
gunicorn==19.0.0
pandas==1.0.3
geopandas==0.9.0
pyproj==2.6.0
dash_core_components==1.13.0
numpy==1.19.3
dash_html_components==1.1.1
dash==1.17.0
dash_extensions==0.0.60
plotly==5.2.1
Flask==1.1.1
PyGithub==1.55


